moved the BaseTaskQueueCallback to the automated package
