+++ src/test/java/teammates/test/automated/FeedbackSubmissionsAdjustmentCallback.java
@@ -8,7 +8,6 @@ import java.util.HashMap;
 import teammates.common.util.Const;
 import teammates.common.util.HttpRequestHelper;
 import teammates.common.util.Const.ParamsNames;
-import teammates.test.cases.BaseTaskQueueCallback;
 
 import com.google.appengine.api.urlfetch.URLFetchServicePb.URLFetchRequest;
